backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Computer Vision and ML
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0

# Utilities
python-dotenv==1.0.0
